+++ example/lib/main.dart
@@ -1,30 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'theme.dart';
 import 'home.dart';
 
+
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-
-  SharedPreferences sp = await SharedPreferences.getInstance();
-  bool isDark = sp.getBool('is_dark');
-
-  ThemeMode initialThemeMode;
-  if (isDark != null) {
-    initialThemeMode = isDark ? ThemeMode.dark : ThemeMode.light;
-  }
-
   runApp(
     EasyDynamicThemeWidget(
-      initialThemeMode: initialThemeMode,
       child: MyApp(),
     ),
   );
 }
 
 class MyApp extends StatelessWidget {
-  final String title = 'Dynamic Theme';
+  final String title = 'EDT - Example';
 
   @override
   Widget build(BuildContext context) {