+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.4
+
+- Refactor and rename switchTheme function to changeTheme.
+- Wrap EasyDynamicThemeWidget build function in a FutureBuilder.
+- Add documentation for changeTheme function.
+
 # 0.0.3
 
 - Minor fix in files format.
@@ -5,7 +11,7 @@
 # 0.0.2
 
 - Minor fix in documentation text.
-- Added a project description.
+- Add project description.
 
 # 0.0.1